+++ finality-provider/service/finality_submitter.go
@@ -26,13 +26,13 @@ var _ types.FinalitySignatureSubmitter = (*DefaultFinalitySubmitter)(nil)
 type PubRandProofListGetterFunc func(startHeight uint64, numPubRand uint64) ([][]byte, error)
 
 type DefaultFinalitySubmitter struct {
-	state               types.FinalityProviderState
-	em                  eotsmanager.EOTSManager
-	consumerCtrl        api.ConsumerController
-	proofListGetterFunc PubRandProofListGetterFunc
-	cfg                 *FinalitySubmitterConfig
-	logger              *zap.Logger
-	metrics             *metrics.FpMetrics
+	State               types.FinalityProviderState
+	Em                  eotsmanager.EOTSManager
+	ConsumerCtrl        api.ConsumerController
+	ProofListGetterFunc PubRandProofListGetterFunc
+	Cfg                 *FinalitySubmitterConfig
+	Logger              *zap.Logger
+	Metrics             *metrics.FpMetrics
 }
 
 type FinalitySubmitterConfig struct {
@@ -61,68 +61,66 @@ func NewDefaultFinalitySubmitter(
 	logger *zap.Logger,
 	metrics *metrics.FpMetrics) *DefaultFinalitySubmitter {
 	return &DefaultFinalitySubmitter{
-		em:                  em,
-		consumerCtrl:        consumerCtrl,
-		proofListGetterFunc: proofListGetterFunc,
-		cfg:                 cfg,
-		logger:              logger.With(zap.String("module", "finality_submitter")),
-		metrics:             metrics,
+		Em:                  em,
+		ConsumerCtrl:        consumerCtrl,
+		ProofListGetterFunc: proofListGetterFunc,
+		Cfg:                 cfg,
+		Logger:              logger.With(zap.String("module", "finality_submitter")),
+		Metrics:             metrics,
 	}
 }
 
-func (ds *DefaultFinalitySubmitter) getBtcPkHex() string {
-	return ds.getBtcPkBIP340().MarshalHex()
+func (ds *DefaultFinalitySubmitter) GetBtcPkHex() string {
+	return ds.GetBtcPkBIP340().MarshalHex()
 }
 
-func (ds *DefaultFinalitySubmitter) getBtcPk() *btcec.PublicKey {
-	return ds.state.GetBtcPk()
+func (ds *DefaultFinalitySubmitter) GetBtcPk() *btcec.PublicKey {
+	return ds.State.GetBtcPk()
 }
 
-func (ds *DefaultFinalitySubmitter) getBtcPkBIP340() *bbntypes.BIP340PubKey {
-	return ds.state.GetBtcPkBIP340()
+func (ds *DefaultFinalitySubmitter) GetBtcPkBIP340() *bbntypes.BIP340PubKey {
+	return ds.State.GetBtcPkBIP340()
 }
 
-func (ds *DefaultFinalitySubmitter) mustSetLastVotedHeight(height uint64) {
-	if err := ds.state.SetLastVotedHeight(height); err != nil {
-		ds.logger.Fatal("failed to update state after finality signature submitted",
-			zap.String("pk", ds.getBtcPkHex()), zap.Uint64("height", height), zap.Error(err))
+func (ds *DefaultFinalitySubmitter) MustSetLastVotedHeight(height uint64) {
+	if err := ds.State.SetLastVotedHeight(height); err != nil {
+		ds.Logger.Fatal("failed to update state after finality signature submitted",
+			zap.String("pk", ds.GetBtcPkHex()), zap.Uint64("height", height), zap.Error(err))
 	}
 }
 
 func (ds *DefaultFinalitySubmitter) mustSetStatus(s proto.FinalityProviderStatus) {
-	if err := ds.state.SetStatus(s); err != nil {
-		ds.logger.Fatal("failed to set finality-provider status",
-			zap.String("pk", ds.getBtcPkHex()), zap.String("status", s.String()))
+	if err := ds.State.SetStatus(s); err != nil {
+		ds.Logger.Fatal("failed to set finality-provider status",
+			zap.String("pk", ds.GetBtcPkHex()), zap.String("status", s.String()))
 	}
 }
 
 // InitState sets the finality provider state store.
 func (ds *DefaultFinalitySubmitter) InitState(state types.FinalityProviderState) error {
-	if ds.state != nil {
+	if ds.State != nil {
 		return fmt.Errorf("finality provider state is already set")
 	}
 
-	ds.state = state
+	ds.State = state
 
 	return nil
 }
 
-// filterBlocksForVoting filters blocks based on the finality provider's voting power and height criteria for submission.
-// It returns a slice of blocks eligible for voting and an error if any issues are encountered during processing.
-// It also updates the fp instance status according to the block's voting power
-func (ds *DefaultFinalitySubmitter) filterBlocksForVoting(ctx context.Context, blocks []types.BlockDescription) ([]types.BlockDescription, error) {
+// FilterBlocksForVoting filters blocks based on the finality provider's voting power and height criteria for submission, returning a slice of blocks eligible for voting and an error if any issues are encountered during processing. It also updates the finality provider instance status according to the block's voting power.
+func (ds *DefaultFinalitySubmitter) FilterBlocksForVoting(ctx context.Context, blocks []types.BlockDescription) ([]types.BlockDescription, error) {
 	processedBlocks := make([]types.BlockDescription, 0, len(blocks))
 
 	var hasPower bool
 	var err error
 	for _, b := range blocks {
 		blk := b
-		if blk.GetHeight() <= ds.state.GetLastVotedHeight() {
-			ds.logger.Debug(
+		if blk.GetHeight() <= ds.State.GetLastVotedHeight() {
+			ds.Logger.Debug(
 				"the block height is lower than last processed height",
-				zap.String("pk", ds.getBtcPkHex()),
+				zap.String("pk", ds.GetBtcPkHex()),
 				zap.Uint64("block_height", blk.GetHeight()),
-				zap.Uint64("last_voted_height", ds.state.GetLastVotedHeight()),
+				zap.Uint64("last_voted_height", ds.State.GetLastVotedHeight()),
 			)
 
 			continue
@@ -135,21 +133,21 @@ func (ds *DefaultFinalitySubmitter) filterBlocksForVoting(ctx context.Context, b
 			return nil, fmt.Errorf("failed to get voting power for height %d: %w", blkHeight, err)
 		}
 		if !hasPower {
-			ds.logger.Debug(
+			ds.Logger.Debug(
 				"the finality-provider does not have voting power",
-				zap.String("pk", ds.getBtcPkHex()),
+				zap.String("pk", ds.GetBtcPkHex()),
 				zap.Uint64("block_height", blkHeight),
 			)
 
 			// the finality provider does not have voting power
 			// and it will never will at this block, so continue
-			ds.metrics.IncrementFpTotalBlocksWithoutVotingPower(ds.getBtcPkHex())
+			ds.Metrics.IncrementFpTotalBlocksWithoutVotingPower(ds.GetBtcPkHex())
 
 			continue
 		}
-		ds.logger.Debug(
+		ds.Logger.Debug(
 			"the finality-provider has voting power",
-			zap.String("pk", ds.getBtcPkHex()),
+			zap.String("pk", ds.GetBtcPkHex()),
 			zap.Uint64("block_height", blkHeight),
 		)
 
@@ -157,11 +155,11 @@ func (ds *DefaultFinalitySubmitter) filterBlocksForVoting(ctx context.Context, b
 	}
 
 	// update fp status according to the power for the last block
-	if hasPower && ds.state.GetStatus() != proto.FinalityProviderStatus_ACTIVE {
+	if hasPower && ds.State.GetStatus() != proto.FinalityProviderStatus_ACTIVE {
 		ds.mustSetStatus(proto.FinalityProviderStatus_ACTIVE)
 	}
 
-	if !hasPower && ds.state.GetStatus() == proto.FinalityProviderStatus_ACTIVE {
+	if !hasPower && ds.State.GetStatus() == proto.FinalityProviderStatus_ACTIVE {
 		ds.mustSetStatus(proto.FinalityProviderStatus_INACTIVE)
 	}
 
@@ -177,16 +175,16 @@ func (ds *DefaultFinalitySubmitter) SubmitBatchFinalitySignatures(ctx context.Co
 		return nil, fmt.Errorf("cannot send signatures for empty blocks")
 	}
 
-	blocks, err := ds.filterBlocksForVoting(ctx, blocks)
+	blocks, err := ds.FilterBlocksForVoting(ctx, blocks)
 	if err != nil {
 		return nil, fmt.Errorf("failed to filter blocks for voting: %w", err)
 	}
 
 	if len(blocks) == 0 {
-		ds.logger.Debug(
+		ds.Logger.Debug(
 			"no blocks to vote for after filtering",
-			zap.String("pk", ds.getBtcPkHex()),
-			zap.Uint64("last_voted_height", ds.state.GetLastVotedHeight()),
+			zap.String("pk", ds.GetBtcPkHex()),
+			zap.Uint64("last_voted_height", ds.State.GetLastVotedHeight()),
 		)
 
 		return nil, nil // No blocks to vote for
@@ -199,9 +197,9 @@ func (ds *DefaultFinalitySubmitter) SubmitBatchFinalitySignatures(ctx context.Co
 	for {
 		res, err := ds.submitBatchFinalitySignaturesOnce(ctx, blocks)
 		if err != nil {
-			ds.logger.Debug(
+			ds.Logger.Debug(
 				"failed to submit finality signature to the consumer chain",
-				zap.String("pk", ds.getBtcPkHex()),
+				zap.String("pk", ds.GetBtcPkHex()),
 				zap.Uint32("current_failures", failedCycles),
 				zap.Uint64("target_start_height", blocks[0].GetHeight()),
 				zap.Uint64("target_end_height", targetHeight),
@@ -218,7 +216,7 @@ func (ds *DefaultFinalitySubmitter) SubmitBatchFinalitySignatures(ctx context.Co
 			}
 
 			failedCycles++
-			if failedCycles > ds.cfg.MaxSubmissionRetries {
+			if failedCycles > ds.Cfg.MaxSubmissionRetries {
 				return nil, fmt.Errorf("reached max failed cycles with err: %w", err)
 			}
 		} else {
@@ -227,29 +225,29 @@ func (ds *DefaultFinalitySubmitter) SubmitBatchFinalitySignatures(ctx context.Co
 		}
 
 		// Check if the block is already finalized
-		finalized, err := ds.checkBlockFinalization(ctx, targetHeight)
+		finalized, err := ds.CheckBlockFinalization(ctx, targetHeight)
 		if err != nil {
 			return nil, fmt.Errorf("failed to query block finalization at height %v: %w", targetHeight, err)
 		}
 		if finalized {
-			ds.logger.Debug(
+			ds.Logger.Debug(
 				"the block is already finalized, skip submission",
-				zap.String("pk", ds.getBtcPkHex()),
+				zap.String("pk", ds.GetBtcPkHex()),
 				zap.Uint64("target_height", targetHeight),
 			)
 
-			ds.metrics.IncrementFpTotalFailedVotes(ds.getBtcPkHex())
+			ds.Metrics.IncrementFpTotalFailedVotes(ds.GetBtcPkHex())
 
 			return nil, nil
 		}
 
 		// Wait for the retry interval
 		select {
-		case <-time.After(ds.cfg.SubmissionRetryInterval):
+		case <-time.After(ds.Cfg.SubmissionRetryInterval):
 			// Continue to next retry iteration
 			continue
 		case <-ctx.Done():
-			ds.logger.Debug("the finality-provider instance is closing", zap.String("pk", ds.getBtcPkHex()))
+			ds.Logger.Debug("the finality-provider instance is closing", zap.String("pk", ds.GetBtcPkHex()))
 
 			return nil, ErrFinalityProviderShutDown
 		}
@@ -280,7 +278,7 @@ func (ds *DefaultFinalitySubmitter) submitBatchFinalitySignaturesOnce(ctx contex
 		prList = append(prList, pr[0])
 
 		// Get proof for this specific height
-		proofs, err := ds.proofListGetterFunc(block.GetHeight(), 1)
+		proofs, err := ds.ProofListGetterFunc(block.GetHeight(), 1)
 		if err != nil {
 			return nil, fmt.Errorf("failed to get public randomness inclusion proof for height %d: %w\nplease recover the randomness proof from db", block.GetHeight(), err)
 		}
@@ -298,13 +296,13 @@ func (ds *DefaultFinalitySubmitter) submitBatchFinalitySignaturesOnce(ctx contex
 
 	// Process each block and collect only valid items
 	for i, b := range blocks {
-		eotsSig, err := ds.signFinalitySig(b)
+		eotsSig, err := ds.SignFinalitySig(b)
 		if err != nil {
 			if !errors.Is(err, ErrFailedPrecondition) {
 				return nil, err
 			}
 			// Skip this block if we encounter FailedPrecondition
-			ds.logger.Warn("encountered FailedPrecondition error, skipping block",
+			ds.Logger.Warn("encountered FailedPrecondition error, skipping block",
 				zap.Uint64("height", b.GetHeight()),
 				zap.String("hash", hex.EncodeToString(b.GetHash())),
 				zap.Error(err))
@@ -321,14 +319,14 @@ func (ds *DefaultFinalitySubmitter) submitBatchFinalitySignaturesOnce(ctx contex
 
 	// If all blocks were skipped, return early
 	if len(validBlocks) == 0 {
-		ds.logger.Info("all blocks were skipped due to double sign errors")
+		ds.Logger.Info("all blocks were skipped due to double sign errors")
 
 		return nil, nil
 	}
 
 	// send finality signature to the consumer chain
-	res, err := ds.consumerCtrl.SubmitBatchFinalitySigs(ctx, api.NewSubmitBatchFinalitySigsRequest(
-		ds.getBtcPk(),
+	res, err := ds.ConsumerCtrl.SubmitBatchFinalitySigs(ctx, api.NewSubmitBatchFinalitySigsRequest(
+		ds.GetBtcPk(),
 		validBlocks,
 		validPrList,
 		validProofList,
@@ -348,19 +346,19 @@ func (ds *DefaultFinalitySubmitter) submitBatchFinalitySignaturesOnce(ctx contex
 
 	// update the metrics with voted blocks
 	for _, b := range validBlocks {
-		ds.metrics.RecordFpVotedHeight(ds.getBtcPkHex(), b.GetHeight())
+		ds.Metrics.RecordFpVotedHeight(ds.GetBtcPkHex(), b.GetHeight())
 	}
 
 	// update state with the highest height of this batch
 	highBlock := blocks[len(blocks)-1]
-	ds.mustSetLastVotedHeight(highBlock.GetHeight())
+	ds.MustSetLastVotedHeight(highBlock.GetHeight())
 
 	return res, nil
 }
 
-// checkBlockFinalization checks if a block at given height is finalized
-func (ds *DefaultFinalitySubmitter) checkBlockFinalization(ctx context.Context, height uint64) (bool, error) {
-	b, err := ds.consumerCtrl.QueryBlock(ctx, height)
+// CheckBlockFinalization checks if a block at given height is finalized.
+func (ds *DefaultFinalitySubmitter) CheckBlockFinalization(ctx context.Context, height uint64) (bool, error) {
+	b, err := ds.ConsumerCtrl.QueryBlock(ctx, height)
 	if err != nil {
 		return false, fmt.Errorf("failed to query block at height %d: %w", height, err)
 	}
@@ -368,17 +366,17 @@ func (ds *DefaultFinalitySubmitter) checkBlockFinalization(ctx context.Context, 
 	return b.IsFinalized(), nil
 }
 
-func (ds *DefaultFinalitySubmitter) signFinalitySig(b types.BlockDescription) (*bbntypes.SchnorrEOTSSig, error) {
+func (ds *DefaultFinalitySubmitter) SignFinalitySig(b types.BlockDescription) (*bbntypes.SchnorrEOTSSig, error) {
 	// build proper finality signature request
 	var msgToSign []byte
-	if b.GetHeight() >= ds.cfg.ContextSigningHeight {
-		signCtx := ds.consumerCtrl.GetFpFinVoteContext()
+	if b.GetHeight() >= ds.Cfg.ContextSigningHeight {
+		signCtx := ds.ConsumerCtrl.GetFpFinVoteContext()
 		msgToSign = b.MsgToSign(signCtx)
 	} else {
 		msgToSign = b.MsgToSign("")
 	}
 
-	sig, err := ds.em.SignEOTS(ds.getBtcPkBIP340().MustMarshal(), ds.state.GetChainID(), msgToSign, b.GetHeight())
+	sig, err := ds.Em.SignEOTS(ds.GetBtcPkBIP340().MustMarshal(), ds.State.GetChainID(), msgToSign, b.GetHeight())
 	if err != nil {
 		if strings.Contains(err.Error(), failedPreconditionErrStr) {
 			return nil, ErrFailedPrecondition
@@ -391,9 +389,9 @@ func (ds *DefaultFinalitySubmitter) signFinalitySig(b types.BlockDescription) (*
 }
 
 func (ds *DefaultFinalitySubmitter) GetPubRandList(startHeight uint64, numPubRand uint32) ([]*btcec.FieldVal, error) {
-	pubRandList, err := ds.em.CreateRandomnessPairList(
-		ds.getBtcPkBIP340().MustMarshal(),
-		ds.state.GetChainID(),
+	pubRandList, err := ds.Em.CreateRandomnessPairList(
+		ds.GetBtcPkBIP340().MustMarshal(),
+		ds.State.GetChainID(),
 		startHeight,
 		numPubRand,
 	)
@@ -411,8 +409,8 @@ func (ds *DefaultFinalitySubmitter) getVotingPowerWithRetry(ctx context.Context,
 	)
 
 	if err := retry.Do(func() error {
-		hasPower, err = ds.consumerCtrl.QueryFinalityProviderHasPower(ctx, api.NewQueryFinalityProviderHasPowerRequest(
-			ds.getBtcPk(),
+		hasPower, err = ds.ConsumerCtrl.QueryFinalityProviderHasPower(ctx, api.NewQueryFinalityProviderHasPowerRequest(
+			ds.GetBtcPk(),
 			height,
 		))
 		if err != nil {
@@ -421,7 +419,7 @@ func (ds *DefaultFinalitySubmitter) getVotingPowerWithRetry(ctx context.Context,
 
 		return nil
 	}, RtyAtt, RtyDel, RtyErr, retry.OnRetry(func(n uint, err error) {
-		ds.logger.Debug(
+		ds.Logger.Debug(
 			"failed to query the voting power",
 			zap.Uint("attempt", n+1),
 			zap.Uint("max_attempts", RtyAttNum),