+++ CHANGELOG.md
@@ -86,6 +86,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
   init fp metrics
 * [#592](https://github.com/babylonlabs-io/finality-provider/pull/592)
   chore(rollup): add unit tests for sparse pub randomness
+* [#593](https://github.com/babylonlabs-io/finality-provider/pull/593)
+  rollup: sparse pub rand commit leftovers
 * [#594](https://github.com/babylonlabs-io/finality-provider/pull/594)
   chore(rollup): remove allowlist requirement for starting rollup FP